refactor(location-manager): modularize vertex content service
