+++ apps/location-manager/packages/python-alt-text/package.json
@@ -6,7 +6,7 @@
     "dev": "node scripts/show-commands.cjs && pnpm run py:ensure && .venv/bin/python -m uvicorn app:app --host 0.0.0.0 --port 8642 --reload",
     "dev:clean": "lsof -ti:8642 | xargs kill -9 2>/dev/null || true; UV_CACHE_DIR=../../../../.uv-cache uv venv --python \"$(node scripts/resolve-python.cjs)\" --clear .venv && UV_CACHE_DIR=../../../../.uv-cache uv pip install --python .venv/bin/python -r requirements.txt && node scripts/show-commands.cjs && .venv/bin/python -m uvicorn app:app --host 0.0.0.0 --port 8642 --reload",
     "build": "echo 'No build step for Python'",
-    "lint": "python3 -m py_compile app.py",
+    "lint": "python3 -m py_compile app.py generation.py grounding.py models.py prompts.py vertex_runtime.py",
     "test": "pnpm run py:ensure && .venv/bin/python -m unittest -q test_app",
     "py:dev": ".venv/bin/python app.py",
     "py:ensure": "UV_CACHE_DIR=../../../../.uv-cache uv venv --python \"$(node scripts/resolve-python.cjs)\" --allow-existing .venv && (.venv/bin/python -c \"import uvicorn, vertexai, google.genai\" >/dev/null 2>&1 || UV_CACHE_DIR=../../../../.uv-cache uv pip install --python .venv/bin/python -r requirements.txt)",
